Flytt tester for søknadService og sjekk om søknader har dokumentasjonsbehov
