+++ pom.xml
@@ -25,7 +25,7 @@
         <felles.version>1.20210712123123_4ab21f1</felles.version>
         <prosessering.version>1.20210617141104_8f0760d</prosessering.version>
         <confluent.version>6.2.0</confluent.version>
-        <kontrakter.version>2.0_20210803091709_b7863e5</kontrakter.version>
+        <kontrakter.version>2.0_20210805110142_99e2341</kontrakter.version>
         <brukernotifikasjon-skjemas.version>1.2020.11.16-15.17-5f7495d2ba72</brukernotifikasjon-skjemas.version>
         <token-validation.version>1.3.8</token-validation.version>
         <spring-cloud.version>3.0.3</spring-cloud.version>
