+++ src/test/kotlin/no/nav/familie/ef/mottak/service/SøknadServiceImplTest.kt
@@ -1,188 +1,50 @@
-package no.nav.familie.ef.mottak.service
+package no.nav.familie.ef.mottak.no.nav.familie.ef.mottak.service
 
-import no.nav.familie.ef.mottak.IntegrasjonSpringRunnerTest
-import no.nav.familie.ef.mottak.api.ApiFeil
-import no.nav.familie.ef.mottak.service.Testdata.skjemaForArbeidssøker
-import no.nav.familie.ef.mottak.service.Testdata.søknadBarnetilsyn
-import no.nav.familie.ef.mottak.service.Testdata.søknadOvergangsstønad
-import no.nav.familie.ef.mottak.service.Testdata.søknadSkolepenger
-import no.nav.familie.ef.mottak.service.Testdata.vedlegg
-import no.nav.familie.kontrakter.ef.søknad.Dokumentasjonsbehov
-import no.nav.familie.kontrakter.ef.søknad.Fødselsnummer
-import no.nav.familie.kontrakter.ef.søknad.SøknadMedVedlegg
-import no.nav.familie.kontrakter.ef.søknad.SøknadType
-import no.nav.familie.kontrakter.ef.søknad.Søknadsfelt
-import no.nav.familie.util.FnrGenerator
+import io.mockk.every
+import io.mockk.mockk
+import no.nav.familie.ef.mottak.mapper.SøknadMapper
+import no.nav.familie.ef.mottak.repository.DokumentasjonsbehovRepository
+import no.nav.familie.ef.mottak.repository.SøknadRepository
+import no.nav.familie.ef.mottak.repository.domain.Dokumentasjonsbehov
+import no.nav.familie.ef.mottak.service.SøknadServiceImpl
+import no.nav.familie.ef.mottak.service.Testdata
+import no.nav.familie.kontrakter.ef.søknad.Dokumentasjonsbehov as DokumentasjonsbehovKontrakter
+import no.nav.familie.kontrakter.felles.objectMapper
 import org.assertj.core.api.Assertions.assertThat
-import org.assertj.core.api.Assertions.catchThrowable
 import org.junit.jupiter.api.Test
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.test.context.ActiveProfiles
-import java.time.LocalDate
+import org.springframework.data.repository.findByIdOrNull
 import java.util.UUID
 
-@ActiveProfiles("local")
-internal class SøknadServiceImplTest : IntegrasjonSpringRunnerTest() {
+class SøknadServiceImplTest {
 
-    @Autowired(required = true) lateinit var søknadService: SøknadService
-
-    @Test
-    internal fun `lagre skjema for arbeidssøker`() {
-        val kvittering = søknadService.motta(skjemaForArbeidssøker)
-        val søknad = søknadService.get(kvittering.id)
-        assertThat(søknad).isNotNull
-    }
+    val søknadRepository = mockk<SøknadRepository>()
+    val dokumentasjonsbehovRepository = mockk<DokumentasjonsbehovRepository>()
+    val søknadService = SøknadServiceImpl(søknadRepository, mockk(), dokumentasjonsbehovRepository, mockk())
 
     @Test
-    internal fun `lagre skjema for søknad`() {
-        val kvittering = søknadService.mottaOvergangsstønad(SøknadMedVedlegg(søknadOvergangsstønad, emptyList()), emptyMap())
-        val søknad = søknadService.get(kvittering.id)
-        assertThat(søknad).isNotNull
-    }
-
-    @Test
-    internal fun `lagre skjema for søknad overgangsstønad med vedlegg`() {
-        val kvittering = søknadService.mottaOvergangsstønad(SøknadMedVedlegg(søknadOvergangsstønad, vedlegg),
-                                                            vedlegg.map { it.id to it.navn.toByteArray() }.toMap())
-        val søknad = søknadService.get(kvittering.id)
-        assertThat(søknad).isNotNull
-    }
-
-
-    @Test
-    internal fun `lagre skjema for skolepenger`() {
-        val kvittering = søknadService.mottaSkolepenger(SøknadMedVedlegg(søknadSkolepenger, emptyList()), emptyMap())
-        val søknad = søknadService.get(kvittering.id)
-        assertThat(søknad).isNotNull
-    }
-
-    @Test
-    internal fun `lagre skjema for søknad skolepenger med vedlegg`() {
-        val kvittering = søknadService.mottaSkolepenger(SøknadMedVedlegg(søknadSkolepenger, vedlegg),
-                                                        vedlegg.map { it.id to it.navn.toByteArray() }.toMap())
-        val søknad = søknadService.get(kvittering.id)
-        assertThat(søknad).isNotNull
-    }
-
-    @Test
-    internal fun `skal kunne behandle sak med barn under 6 mnd i ny løsning`() {
-        val barnFødtIDag = listOf(søknadOvergangsstønad.barn.verdi.first().copy(
-                fødselTermindato = Søknadsfelt("Termindato", LocalDate.now())
-        ))
-        val søknadMedBarnFødtIDag = søknadOvergangsstønad.copy(
-                barn = Søknadsfelt("Barn", barnFødtIDag)
+    internal fun `hentDokumentasjonsbehovforPerson fungerer for overgangsstønad, barnetilsyn og skolepenger`() {
+        val fnr = "12345678"
+        val søknader = listOf(
+                SøknadMapper.fromDto(Testdata.søknadOvergangsstønad, false),
+                SøknadMapper.fromDto(Testdata.søknadBarnetilsyn, false),
+                SøknadMapper.fromDto(Testdata.søknadSkolepenger, false),
+                SøknadMapper.fromDto(Testdata.skjemaForArbeidssøker),
         )
-        val kvittering = søknadService.mottaOvergangsstønad(SøknadMedVedlegg(søknadMedBarnFødtIDag, vedlegg),
-                                                            vedlegg.map { it.id to it.navn.toByteArray() }.toMap())
+        val forventetDokumentasjonsbehov = listOf(DokumentasjonsbehovKontrakter(
+                "test",
+                UUID.randomUUID()
+                        .toString(),
+                false))
 
-        val søknad = søknadService.get(kvittering.id)
-        assertThat(søknad).isNotNull
-        assertThat(søknad.behandleINySaksbehandling).isTrue
+        every { søknadRepository.findAllByFnr(fnr) } returns søknader
+
+        every { dokumentasjonsbehovRepository.findByIdOrNull(any()) } returns Dokumentasjonsbehov("123",
+                                                                                                  objectMapper.writeValueAsString(
+                                                                                                          forventetDokumentasjonsbehov))
+
+        every { søknadRepository.findByIdOrNull(any()) } returns SøknadMapper.fromDto(Testdata.søknadOvergangsstønad, false)
+
+
+        assertThat(søknadService.hentDokumentasjonsbehovForPerson(fnr)).hasSize(3)
     }
-
-    @Test
-    internal fun `skal kunne behandle sak med et barn under 6 mnd og et over i ny løsning`() {
-        val barnFødtIDag = søknadOvergangsstønad.barn.verdi.first().copy(
-                fødselTermindato = Søknadsfelt("Termindato", LocalDate.now())
-        )
-
-        val barnFødtforLengeSiden = søknadOvergangsstønad.barn.verdi.first().copy(
-                fødselTermindato = Søknadsfelt("Termindato", LocalDate.now().minusYears(3))
-        )
-        val søknadMedBarnFødtIDag = søknadOvergangsstønad.copy(
-                barn = Søknadsfelt("Barn", listOf(barnFødtforLengeSiden, barnFødtIDag))
-        )
-        val kvittering = søknadService.mottaOvergangsstønad(SøknadMedVedlegg(søknadMedBarnFødtIDag, vedlegg),
-                                                            vedlegg.map { it.id to it.navn.toByteArray() }.toMap())
-
-        val søknad = søknadService.get(kvittering.id)
-        assertThat(søknad).isNotNull
-        assertThat(søknad.behandleINySaksbehandling).isTrue
-    }
-
-    @Test
-    internal fun `skal ikke kunne behandle sak med barn over 6 mnd i ny løsning`() {
-        val barn7mnd = listOf(søknadOvergangsstønad.barn.verdi.first().copy(
-                fødselTermindato = Søknadsfelt("Termindato", LocalDate.now().minusMonths(7))
-        ))
-        val søknadMedBarn7mnd = søknadOvergangsstønad.copy(
-                barn = Søknadsfelt("Barn", barn7mnd)
-        )
-        val kvittering = søknadService.mottaOvergangsstønad(SøknadMedVedlegg(søknadMedBarn7mnd, vedlegg),
-                                                            vedlegg.map { it.id to it.navn.toByteArray() }.toMap())
-        val søknad = søknadService.get(kvittering.id)
-        assertThat(søknad).isNotNull
-        assertThat(søknad.behandleINySaksbehandling).isFalse
-    }
-
-    @Test
-    internal fun `skal ikke behandle sak med barn over 6 mnd i ny løsning gitt ident `() {
-        val fødselsdato = LocalDate.now().minusMonths(7)
-        val ident = FnrGenerator.generer(fødselsdato.year, fødselsdato.month.value, fødselsdato.dayOfMonth)
-
-        val barn1mnd = listOf(søknadOvergangsstønad.barn.verdi.first().copy(
-                fødselsnummer = Søknadsfelt("Fødselsnummer", Fødselsnummer(ident))
-        )
-        )
-        val søknadMedBarn1mnd = søknadOvergangsstønad.copy(
-                barn = Søknadsfelt("Barn", barn1mnd)
-        )
-        val kvittering = søknadService.mottaOvergangsstønad(SøknadMedVedlegg(søknadMedBarn1mnd, vedlegg),
-                                                            vedlegg.map { it.id to it.navn.toByteArray() }.toMap())
-        val søknad = søknadService.get(kvittering.id)
-        assertThat(søknad).isNotNull
-        assertThat(søknad.behandleINySaksbehandling).isFalse
-    }
-
-
-    @Test
-    internal fun `skal kunne behandle sak med barn under 6 mnd i ny løsning gitt ident `() {
-        val fødselsdato = LocalDate.now().minusMonths(5)
-        val ident = FnrGenerator.generer(fødselsdato.year, fødselsdato.month.value, fødselsdato.dayOfMonth)
-
-        val barn1mnd = listOf(søknadOvergangsstønad.barn.verdi.first().copy(
-                fødselsnummer = Søknadsfelt("Fødselsnummer", Fødselsnummer(ident))
-        )
-        )
-        val søknadMedBarn1mnd = søknadOvergangsstønad.copy(
-                barn = Søknadsfelt("Barn", barn1mnd)
-        )
-        val kvittering = søknadService.mottaOvergangsstønad(SøknadMedVedlegg(søknadMedBarn1mnd, vedlegg),
-                                                            vedlegg.map { it.id to it.navn.toByteArray() }.toMap())
-        val søknad = søknadService.get(kvittering.id)
-        assertThat(søknad).isNotNull
-        assertThat(søknad.behandleINySaksbehandling).isTrue
-    }
-
-
-    @Test
-    internal fun `lagre skjema for søknad barnetilsyn`() {
-        val kvittering = søknadService.mottaBarnetilsyn(SøknadMedVedlegg(søknadBarnetilsyn, emptyList()), emptyMap())
-        val søknad = søknadService.get(kvittering.id)
-        assertThat(søknad).isNotNull
-    }
-
-    @Test
-    internal fun `hent dokumentasjonsbehov til søknad`() {
-        val dokumentasjonsbehov = listOf(Dokumentasjonsbehov("label", "id", false, emptyList()))
-        val søknad = søknadOvergangsstønad
-        val kvittering =
-                søknadService.mottaOvergangsstønad(SøknadMedVedlegg(søknad, emptyList(), dokumentasjonsbehov),
-                                                   emptyMap())
-        val dokumentasjonsbehovDto = søknadService.hentDokumentasjonsbehovForSøknad(UUID.fromString(kvittering.id))
-
-        assertThat(dokumentasjonsbehovDto.personIdent).isEqualTo(søknad.personalia.verdi.fødselsnummer.verdi.verdi)
-        assertThat(dokumentasjonsbehovDto.dokumentasjonsbehov).hasSize(1)
-        assertThat(dokumentasjonsbehovDto.dokumentasjonsbehov[0].harSendtInn).isFalse
-        assertThat(dokumentasjonsbehovDto.dokumentasjonsbehov[0].opplastedeVedlegg).isEmpty()
-        assertThat(dokumentasjonsbehovDto.søknadType).isEqualTo(SøknadType.OVERGANGSSTØNAD)
-    }
-
-    @Test
-    internal fun `hent dokumentasjonsbehov til søknad feiler når søknadId ikke finnes`() {
-        assertThat(catchThrowable { søknadService.hentDokumentasjonsbehovForSøknad(UUID.randomUUID()) })
-                .hasMessageContaining("Fant ikke dokumentasjonsbehov for søknad ")
-                .isInstanceOf(ApiFeil::class.java)
-    }
-
 }
